refactor & make reusable user page components
