+++ packages/client/src/app/_components/index.ts
@@ -6,5 +6,5 @@ export * from './transactions'
 export * from './groups'
 export * from './top-bar'
 export * from './bottom-bar'
-export * from './friends'
+export * from './users'
 export * from './atoms'
