+++ packages/client/src/app/_components/users/index.ts
@@ -1,3 +1,5 @@
 export * from './friends-list'
 export * from './friend-page'
 export * from './user-profile'
+export * from './name-field'
+export * from './user-avatar'
