+++ packages/common/src/types/index.ts
@@ -5,10 +5,16 @@ export * from './old'
 
 export type Function<Props = void, Result = void> = (props: Props) => Result
 
-export type Undefined<T> = Partial<Record<keyof T, undefined>>
 export type Defined<T> = Required<{
   [P in keyof T]-?: Exclude<T[P], undefined>
 }>
+export type Undefined<T> = Partial<Record<keyof T, undefined>>
+
+// either all are defined or all are undefined
+export type Optional<T> = T extends object ? Undefined<T> | {
+  [P in keyof T]: T[P]
+} : Partial<T>
+
 
 export type Loadable<T = {}> = ({
   isLoading: true
